move slug meta code to own component
